feat: add delegation heuristics for orchestrator solution path selection
New DelegationHeuristics struct with speed/cost/quality multipliers,
capability areas, delegate-when/dont-delegate-when rules, and
rule-of-thumb - inspired by oh-my-opencode-slim's orchestrator pattern.

- config.DelegationHeuristics: structured metadata for agent routing
  decisions (speed, cost, quality multipliers + capability areas)
- BuildAgentCatalog(): dynamically generates AGENT CATALOG with tool
  groups, skills, stats, and routing rules for the orchestrator prompt
- summarizeToolGroups(): condenses tool lists into capability categories
  (Web search, Graph and memory, Management, Memory ops, Coordination)
- Annotated 4 delegatable agents (researcher, codebase, agent-creator,
  schema-designer) with full heuristics
- dianeDefaultSystemPrompt(): injects dynamic catalog into system prompt
- Sync logic: pushes delegation into AgentDefinition.Config map on
  seed/update (works with MP's existing Config field)
- Support for user-defined agents via diane.yml delegation field
- MP Issue #199: list_available_agents needs to expose Config

+++ server/internal/mcpproxy/client.go
@@ -185,7 +185,7 @@ func (c *MCPClient) sendRequest(method string, params json.RawMessage) (json.Raw
 	// Create response channel
 	respCh := make(chan MCPResponse, 1)
 	c.pendingMu.Lock()
-	c.pending[reqID] = respCh
+	c.pending[float64(reqID)] = respCh // Use float64 to match JSON decoder's default number type
 	c.pendingMu.Unlock()
 
 	// Send request
@@ -202,7 +202,7 @@ func (c *MCPClient) sendRequest(method string, params json.RawMessage) (json.Raw
 
 	if err != nil {
 		c.pendingMu.Lock()
-		delete(c.pending, reqID)
+		delete(c.pending, float64(reqID))
 		c.pendingMu.Unlock()
 		return nil, fmt.Errorf("failed to send %s: %w", method, err)
 	}
